+++ Engine/Asset/DocumentAssetLoader.h
@@ -8,6 +8,6 @@ class DocumentAssetLoader : public DocumentResourceLoader
 {
 public:
   virtual Any LoadResource(czstr path, DocumentResourceLoaderCallback load_callback, void * user_ptr) override;
-  virtual void UnloadResource(uint64_t path_hash, Any && load_data) override;
+  virtual void UnloadResource(uint32_t path_hash, Any && load_data) override;
 };
