+++ Foundation/Document/DocumentCompiler.h
@@ -22,7 +22,7 @@ public:
 
   NullOptPtr<Document> CreateNewDocument(czstr path);
   NotNullPtr<Document> GetDocument(czstr path);
-  NullOptPtr<Document> GetDocument(uint64_t path_hash);
+  NullOptPtr<Document> GetDocument(uint32_t path_hash);
 
   void ReloadDocument(czstr path);
 
@@ -30,14 +30,14 @@ private:
 
   friend class Document;
 
-  void HandleDocumentLoaded(uint64_t file_id, Optional<Buffer> && buffer, std::chrono::system_clock::time_point last_modified_time);
-  void UnloadDocument(uint64_t file_id);
+  void HandleDocumentLoaded(uint32_t file_id, Optional<Buffer> && buffer, std::chrono::system_clock::time_point last_modified_time);
+  void UnloadDocument(uint32_t file_id);
 
 private:
 
   NotNullPtr<DocumentLoader> m_Loader;
 
-  std::unordered_map<uint64_t, std::unique_ptr<Document>> m_Documents;
+  std::unordered_map<uint32_t, std::unique_ptr<Document>> m_Documents;
 };
